fix: compose build, add shell if

+++ Inception/srcs/requirements/mariadb/conf/start_mariadb.sh
@@ -1,62 +1,65 @@
 #!/bin/sh
 
 # mysql 데이터 디렉토리 초기화
-mysql_install_db --user=mysql --datadir=/var/lib/mysql/
 
-mkdir -p /run/mysqld/
-chown -R mysql:mysql /var/lib/mysql/ /run/mysqld/
+if [ ! -d /var/lib/mysql/$MARIADB_DATABASE ]; then
+	mysql_install_db --user=mysql --datadir=/var/lib/mysql/
 
-chmod 777 /var/lib/mysql
+	mkdir -p /run/mysqld/
+	chown -R mysql:mysql /var/lib/mysql/ /run/mysqld/
 
-# BackGround로 실행한다.
-/usr/bin/mysqld_safe --user=mysql --datadir=/var/lib/mysql/ &
+	chmod 777 /var/lib/mysql
 
-sleep 1
+	# BackGround로 실행한다.
+	/usr/bin/mysqld_safe --user=mysql --datadir=/var/lib/mysql/ &
 
-
-# ===================================================
-
-# wordpress database 생성 SQL문
-
-# database 생성
-mysql -e "create database $MY_MYSQL_DATABASE;"
-
-# database 유저 생성
-mysql -e "create user '$MY_MYSQL_USER'@'%' identified by '$MY_MYSQL_PASSWORD';"
-
-# database 권한
-mysql -e "GRANT SELECT, INSERT, UPDATE, DELETE, CREATE, DROP, INDEX, ALTER, CREATE TEMPORARY TABLES, CREATE VIEW, EVENT, TRIGGER, SHOW VIEW, CREATE ROUTINE, ALTER ROUTINE, EXECUTE ON $MY_MYSQL_DATABASE.* TO '$MY_MYSQL_USER'@'%';"
-
-# 적용
-mysql -e "FLUSH PRIVILEGES;"
-
-# =====================================================
-
-# mysql_secure_installation 설정이 담긴 sql쿼리문
-
-# -- mariadb 10.5부터 native, socket 인증방식 동시사용가능, native인증으로 고정 및 root 패스워드 변경
-mysql -e "ALTER USER root@localhost IDENTIFIED VIA mysql_native_password USING PASSWORD('$MY_MYSQL_ROOT_PASSWORD'); \
-DELETE FROM mysql.user WHERE User=''; \
-DROP DATABASE IF EXISTS test; \
-DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%'; \
-FLUSH PRIVILEGES;"
+	sleep 1
 
 
-# -- '@localhost' 익명연결 허용 삭제
-#mysql -e "DELETE FROM mysql.user WHERE User='';"
+	# ===================================================
 
-# -- testdb 삭제
-#mysql -e "DROP DATABASE IF EXISTS test;"
-#mysql -e "DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%';"
+	# wordpress database 생성 SQL문
 
-# -- 적용
-#mysql -e "FLUSH PRIVILEGES;"
+	# database 생성
+	mysql -e "create database $MARIADB_DATABASE;"
+
+	# database 유저 생성
+	mysql -e "create user '$MARIADB_USER'@'%' identified by '$MARIADB_PASSWORD';"
+
+	# database 권한
+	mysql -e "GRANT SELECT, INSERT, UPDATE, DELETE, CREATE, DROP, INDEX, ALTER, CREATE TEMPORARY TABLES, CREATE VIEW, EVENT, TRIGGER, SHOW VIEW, CREATE ROUTINE, ALTER ROUTINE, EXECUTE ON $MARIADB_DATABASE.* TO '$MARIADB_USER'@'%';"
+
+	# 적용
+	mysql -e "FLUSH PRIVILEGES;"
+
+	# =====================================================
+
+	# mysql_secure_installation 설정이 담긴 sql쿼리문
+
+	# -- mariadb 10.5부터 native, socket 인증방식 동시사용가능, native인증으로 고정 및 root 패스워드 변경
+	mysql -e "ALTER USER root@localhost IDENTIFIED VIA mysql_native_password USING PASSWORD('$MARIADB_ROOT_PASSWORD'); \
+	DELETE FROM mysql.user WHERE User=''; \
+	DROP DATABASE IF EXISTS test; \
+	DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%'; \
+	FLUSH PRIVILEGES;"
 
 
-# mysqld_safe BackGround 종료
-mysqladmin -u root --password=$MY_MYSQL_ROOT_PASSWORD shutdown
+	# -- '@localhost' 익명연결 허용 삭제
+	#mysql -e "DELETE FROM mysql.user WHERE User='';"
 
-sleep 1
+	# -- testdb 삭제
+	#mysql -e "DROP DATABASE IF EXISTS test;"
+	#mysql -e "DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%';"
 
+	# -- 적용
+	#mysql -e "FLUSH PRIVILEGES;"
+
+
+	# mysqld_safe BackGround 종료
+	mysqladmin -u root --password=$MARIADB_ROOT_PASSWORD shutdown
+
+	sleep 1
+fi
 # ForeGround로 실행한다.
-/usr/bin/mysqld_safe
+/usr/bin/mysqld_safe --user=mysql --datadir=/var/lib/mysql/
+
